VoxCeleb1/requirements.txt
appdirs==1.4.4
audioread==3.0.0
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
fonttools==4.39.4
importlib-resources==5.12.0
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lazy-loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.1rc1
matplotlib==3.7.1
msgpack==1.0.5
msgpack-numpy==0.4.8
nnaudio==0.3.1
numba==0.57.0
pooch==1.6.0
primepy==1.3
pydevd==2.9.6
pyparsing==3.0.9
pyspikes==0.2.9
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
spikingjelly==0.0.0.0.14
thop==0.1.1-2209072238
threadpoolctl==3.1.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio-augmentations==0.2.4
wavaugment==0.2
